Remove errant method
This method shouldn’t exist because we can’t properly create an instance of `IldcpRequestPacket` without also encoding the `data` of the packet. However, due to the project structure, the Codecs cannot be used as a dependency here. Thus, this method should exist elsewhere (such as in a Connector code-base).

Signed-off-by: sappenin <sappenin@gmail.com>